Fix types

+++ Amg-ingressos-aqui-eventos-api/Model/Querys/GetEvents.cs
@@ -129,7 +129,7 @@ namespace Amg_ingressos_aqui_eventos_api.Model.Querys
         public long TotalTickets { get; set; }
 
         [JsonProperty("ValueTotal")]
-        public long ValueTotal { get; set; }
+        public float ValueTotal { get; set; }
 
         [JsonProperty("StartDateSales")]
         public DateTime StartDateSales { get; set; }
